content-gen/src/backend/requirements.txt
# Content Generation Solution Accelerator - Python Dependencies

# Web Framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.17.0

# Microsoft Agent Framework
agent-framework==1.0.0b260114  # Pinning this stable version

# Azure SDKs
azure-identity>=1.17.0
azure-cosmos>=4.7.0
azure-storage-blob>=12.22.0
azure-ai-contentsafety>=1.0.0
azure-ai-projects>=1.0.0b5  # Azure AI Foundry SDK (optional, for USE_FOUNDRY=true)

# OpenAI
openai>=1.45.0

# HTTP Client (for Foundry direct API calls)
httpx>=0.27.0

# Data Validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
